handle_orders/requirements.txt
boto3==1.26.151
pandas==2.1.3
yagmail==0.15.293
croniter==2.0.5
pytz
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0